Tokeniser bugs: numeric parsing rewritten, and getint() returning a double (the perils of cut and paste) leading to peculiar roundings of large numbers.

+++ include/tokeniser.h
@@ -109,7 +109,7 @@ public:
             return -9999;
     }
     /// get the value of the next token - must be a float or int
-    float getint()
+    int getint()
     {
         if(curtype==inttoken)
             return val.i;
